--- src/components/OptionPage.module.css
@@ -1,9 +1,0 @@
-.optionPage {
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-}
-
-.optionInput {
-  flex-grow: 1;
-}
