+++ src/components/header/HeaderMenu.module.css
@@ -1,0 +1,69 @@
+.backdrop {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  z-index: 10;
+  background: rgba(0, 0, 0, 0.6);
+}
+
+.menu {
+  position: fixed;
+  right: 0;
+  top: calc(53px + 1vw);
+
+  color: #23006d;
+  font-family: "SF Pro", "Lucida Sans", "Lucida Sans Regular", "Lucida Grande",
+    "Lucida Sans Unicode", Geneva, Verdana, sans-serif;
+  font-size: calc(15px + 0.8vw);
+  background-color: rgb(255, 255, 255);
+  box-shadow: 0 30px 0 rgba(0, 0, 0, 0.5),
+    inset -2px -2px 4px rgba(255, 255, 255, 0.5),
+    inset 2px 2px 4px rgba(170, 170, 204, 0.25),
+    inset 5px 5px 10px rgba(170, 170, 204, 0.5), inset -5px -5px 10px #b3b3b390;
+  border-radius: 10px;
+
+  z-index: 100;
+}
+
+ul.menuList {
+  padding: 10px;
+}
+
+.menuList li {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 10px;
+  border-radius: 5px;
+}
+
+li.changeMusic {
+  padding: 10px;
+  border-radius: 5px;
+  cursor: pointer;
+}
+
+li.changeMusic:hover {
+  padding: 10px;
+  background: #23006d;
+  color: rgb(255, 255, 255);
+  border-radius: 5px;
+}
+
+.on-off {
+  padding: 10px;
+  background: #23006d;
+  color: rgb(255, 255, 255);
+  border-radius: 5px;
+  cursor: pointer;
+}
+
+.on-off:hover {
+  background: #643cba;
+}
+
+.hide {
+  display: none;
+}
