+++ src/components/option-page/OptionPage.module.css
@@ -1,0 +1,9 @@
+.optionPage {
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+}
+
+.optionInput {
+  flex-grow: 1;
+}
