+++ src/components/start-page/StartPage.module.css
@@ -1,0 +1,7 @@
+.start {
+  height: calc(100vh - 8.58vh);
+  display: flex;
+  flex-flow: column wrap;
+  justify-content: center;
+  align-items: center;
+}
